Fix crash and add leading comment for package (#1)
The binary search returns Error when it doesn't find a location for a
path. The unwrap on the return value caused a panic, but in this case,
it should just return None. This caused crashes for files with leading
comments.

Additionally, supports comments leading the package name. They were
being dropped/ignored.